backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-genai
python-multipart
pytest
httpx
slowapi
